chore(meshconfig): fix form
