+++ src/components/inputs/InputField.tsx
@@ -1,30 +1,32 @@
 import { ComponentChild } from "preact";
+import { JSXInternal } from "preact/src/jsx";
 import { FieldValues, Path } from "react-hook-form";
-import { UseFormRegister } from "react-hook-form/dist/types/form";
-import { RegisterOptions } from "react-hook-form/dist/types/validator";
 
 const InputField = <TFieldValues extends FieldValues>({
     id,
     label,
-    register,
-    options,
     error,
+    value,
+    ...inputProps
 }: {
     id: Path<TFieldValues>;
-    label: string | ComponentChild;
-    register?: UseFormRegister<TFieldValues>;
-    options?: RegisterOptions;
+    label?: string | ComponentChild;
     error?: string | ComponentChild;
-}) => {
+} & Partial<
+    Omit<JSXInternal.HTMLAttributes<HTMLInputElement>, "label"> & {
+        defaultValue?: string;
+    }
+>) => {
     return (
-        <div>
-            <label htmlFor={id}>{label}</label>
+        <div className={"flex flex-col w-100"}>
+            {label && <label htmlFor={id}>{label}</label>}
             <input
                 type="text"
                 id={id}
                 data-testid="password-input"
-                {...register(id, { ...options })}
                 className="w-100"
+                value={value}
+                {...inputProps}
             />
             {error && <p class="text-red-500 text-md mt-1">{error}</p>}
         </div>
